feat: signUp함수 바인딩, UserRepository에 CoreDataService 이니셜라이저 추가

+++ LottoDairy/Data/Interfaces/CoreDataGoalAmountEntityPersistenceServiceProtocol.swift
@@ -9,5 +9,5 @@ import Foundation
 import Combine
 
 protocol CoreDataGoalAmountEntityPersistenceServiceProtocol {
-    func saveGoalAmountEntity(_ goalAmount: Int16) -> AnyPublisher<Int16, Error>
+    func saveGoalAmountEntity(_ goalAmount: Int) -> AnyPublisher<Int, Error>
 }
